Internal config values not stored in user config
Previously if the user config (coc-settings.json) did not specify the
location of serve-d and dcd-* binaries, the plugin would use it's
defaults and write them to the user config.

This is problematic as the user config is meant to be portable and
sharable, so it cannot contain absolute paths.

The solution is to use the globalState Memento object provided by coc.